# Add DH_check support to perf tool